# lazy require modules inside subcommands