--- a/src/commands/share.js
+++ b/src/commands/share.js
@@ -1,11 +1,3 @@
-var Dat = require('dat-node')
-var neatLog = require('neat-log')
-var archiveUI = require('../ui/archive')
-var trackArchive = require('../lib/archive')
-var onExit = require('../lib/exit')
-var parseArgs = require('../parse-args')
-var debug = require('debug')('dat')
-
 module.exports = {
   name: 'share',
   command: share,
@@ -38,6 +30,14 @@ module.exports = {
 }
 
 function share (opts) {
+  var Dat = require('dat-node')
+  var neatLog = require('neat-log')
+  var archiveUI = require('../ui/archive')
+  var trackArchive = require('../lib/archive')
+  var onExit = require('../lib/exit')
+  var parseArgs = require('../parse-args')
+  var debug = require('debug')('dat')
+
   if (!opts.dir) {
     opts.dir = parseArgs(opts).dir || process.cwd()
   }