--- a/src/commands/status.js
+++ b/src/commands/status.js
@@ -1,10 +1,3 @@
-var Dat = require('dat-node')
-var neatLog = require('neat-log')
-var statusUI = require('../ui/status')
-var onExit = require('../lib/exit')
-var parseArgs = require('../parse-args')
-var debug = require('debug')('dat')
-
 module.exports = {
   name: 'status',
   command: status,
@@ -17,6 +10,13 @@ module.exports = {
 }
 
 function status (opts) {
+  var Dat = require('dat-node')
+  var neatLog = require('neat-log')
+  var statusUI = require('../ui/status')
+  var onExit = require('../lib/exit')
+  var parseArgs = require('../parse-args')
+  var debug = require('debug')('dat')
+
   debug('dat status')
   if (!opts.dir) {
     opts.dir = parseArgs(opts).dir || process.cwd()