Improves latency, #820 .

Before:

```
dat --help 0.35s user 0.05s system 101% cpu 0.393 total
```

After:

```
dat --help  0.10s user 0.02s system 90% cpu 0.135 total
```